requirements.txt
alembic==1.10.4
asgiref==3.6.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
idna==3.4
ijson==3.2.0.post0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
openpyxl==3.1.2
psycopg2-binary==2.9.6
python-decouple==3.8
requests==2.29.0
SQLAlchemy==2.0.12
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
zipp==3.15.0
